refactor: extract TransportParseError to @livestore/common/leader-thread
Deduplicate TransportParseError and LeaderWorkerTransportError that were
defined identically in both the node and web adapter worker schemas.
Also align the node adapter schema helpers (VoidSchema,
TransferableUint8ArrayFromArrayBuffer) with the web adapter.
