+++ packages/@livestore/adapter-web/src/web-worker/common/worker-schema.ts
@@ -8,10 +8,10 @@ import {
   SyncState,
   UnknownError,
 } from '@livestore/common'
-import { StreamEventsOptionsFields } from '@livestore/common/leader-thread'
+import { LeaderWorkerTransportError, StreamEventsOptionsFields } from '@livestore/common/leader-thread'
 import { EventSequenceNumber, LiveStoreEvent } from '@livestore/common/schema'
 import * as WebmeshWorker from '@livestore/devtools-web-common/worker'
-import { ParseResult, Schema, Transferable, WorkerError } from '@livestore/utils/effect'
+import { Schema, Transferable } from '@livestore/utils/effect'
 
 export const StorageTypeOpfs = Schema.Struct({
   type: Schema.Literal('opfs'),
@@ -54,40 +54,6 @@ const TransferableUint8ArrayFromArrayBuffer = Transferable.schema(
   ),
   (_) => [_.buffer],
 )
-
-const TransportParseErrorEncoded = Schema.Struct({
-  _tag: Schema.Literal('ParseError'),
-  message: Schema.String,
-})
-
-/**
- * Effect's `ParseError` contains schema AST internals that are not stable worker payloads.
- * We only ship the formatted message and reconstruct a real `ParseError` on decode so callers
- * still receive the original transport error type.
- */
-export class TransportParseError extends Schema.transformOrFail(
-  TransportParseErrorEncoded,
-  Schema.instanceOf(ParseResult.ParseError),
-  {
-    strict: true,
-    decode: ({ message }) =>
-      ParseResult.succeed(
-        new ParseResult.ParseError({
-          issue: new ParseResult.Type(Schema.Unknown.ast, undefined, message),
-        }),
-      ),
-    encode: (error) =>
-      ParseResult.succeed(
-        TransportParseErrorEncoded.make({
-          _tag: 'ParseError',
-          message: error.message,
-        }),
-      ),
-  },
-) {}
-
-export const LeaderWorkerTransportError = Schema.Union(WorkerError.WorkerError, TransportParseError)
-export type LeaderWorkerTransportError = typeof LeaderWorkerTransportError.Type
 
 export class LeaderWorkerOuterInitialMessage extends Schema.TaggedRequest<LeaderWorkerOuterInitialMessage>()(
   'InitialMessage',
