+++ packages/@livestore/common/src/leader-thread/mod.ts
@@ -7,3 +7,4 @@ export * from './recreate-db.ts'
 export * as ShutdownChannel from './shutdown-channel.ts'
 export * from './stream-events.ts'
 export * from './types.ts'
+export * from './worker-transport-error.ts'
